Add luarocks as additional installed tool
Luarocks tool was removed in commit:
(b646d99f336cf8104325d29170a5f36361cd3699 "penlight package dependencies broken")

Luarocks tool was used before for installing packages.
After mentioned commit luarocks tool changed to tarantoolctl
and due to no need in it was in the build it was completely
removed from installation build of the images. For now found
that some users still need it. It was decided to return
luarocks tool installed in the image, but not to use it for
packages installations in images builds.

Part of #140