Add ldecnumber installation
According to PR-109 [1] it was decided to add ldecnumber installation.
Depending that 2.x images are lack of decimals support for now checked
Tarantool releases and tags that need to be updated with it. Found that
Tarantool since 2.2 release branches and 2.2.1 release tags are lack of
it, than ldecnumber number installation was added to the earlier tags
and release branches.

Closes #77
Closes #112

[1] - https://github.com/tarantool/docker/pull/109

Co-authored-by: Alexander V. Tikhonov <avtikhon@tarantool.org>